 requirements.txt
# requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0
openai==1.3.7
google-generativeai==0.3.2
sendgrid==6.10.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
discord.py==2.3.2
requests==2.31.0
aiohttp==3.9.1